+++ RELEASE_NOTES.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## Next version
+_Date TBD_
+
+### New features
+* Added direct link to CP activity by CP id
+
 ## Version 0.5.0
 _25 March 2016_